requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==19.3.0
autobahn==22.2.2
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
Click==7.0
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
dbus-python==1.2.16
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2
django-allauth==0.49.0
django-braces==1.15.0
django-online-counter==1.0
django-seed==0.2.2
django-widget-tweaks==1.4.12
entrypoints==0.3
Faker==13.2.0
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
msgpack==1.0.3
netifaces==0.10.4
oauthlib==3.1.0
outcome==1.1.0
pexpect==4.6.0
Pillow==9.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.1.3
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
sos==4.1
sqlparse==0.4.2
ssh-import-id==5.10
supervisor==4.1.0
systemd-python==234
toposort==1.7
trio==0.20.0
trio-websocket==0.9.2
Twisted==18.9.0
txaio==22.2.1
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
wadllib==1.3.3
wsproto==1.1.0
zipp==1.0.0
zope.interface==4.7.1
